Split tests and clean up (#471)
* split tests

* clean up cruft

* Fix NLPD to provide energy readings

* Add BNLD - Bticino dimmable light
