--- .pylintrc
@@ -1,51 +1,0 @@
-[BASIC]
-good-names-rgxs=
-  NA.*
-good-names=on, off
-
-[MESSAGES CONTROL]
-# Reasons disabled:
-# duplicate-code - unavoidable
-# cyclic-import - doesn't test if both import on load
-# unused-argument - generic callbacks and setup methods create a lot of warnings
-# global-statement - used for the on-demand requirement installation
-# too-many-* - are not enforced for the sake of readability
-# too-few-* - same as too-many-*
-# abstract-method - with intro of async there are always methods missing
-# not-an-iterable - https://github.com/PyCQA/pylint/issues/2311
-disable=
-  abstract-method,
-  cyclic-import,
-  duplicate-code,
-  global-statement,
-  inconsistent-return-statements,
-  missing-docstring,
-  too-few-public-methods,
-  too-many-arguments,
-  too-many-branches,
-  too-many-instance-attributes,
-  too-many-lines,
-  too-many-locals,
-  too-many-public-methods,
-  too-many-return-statements,
-  too-many-statements,
-  abstract-method,
-  not-an-iterable,
-  format,
-
-[REPORTS]
-reports=no
-
-[TYPECHECK]
-# For attrs
-ignored-classes=_CountingAttr
-generated-members=botocore.errorfactory
-
-[FORMAT]
-expected-line-ending-format=LF
-
-[EXCEPTIONS]
-overgeneral-exceptions=Exception
-
-[MASTER]
-load-plugins=pylint_pytest
