+++ README.md
@@ -1,5 +1,4 @@
-pyatmo
-======
+# pyatmo
 
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/ambv/black)
 [![GitHub Actions](https://github.com/jabesq/pyatmo/workflows/Python%20package/badge.svg)](https://github.com/jabesq/pyatmo/actions?workflow=Python+package)
@@ -11,16 +10,14 @@ pyatmo
 >
 > I apologize for any inconvenience this may cause, and I sincerely hope to have the capacity to allocate more time to this repository in the near future. Your understanding is greatly appreciated.
 
-***
-
+---
 
 Simple API to access Netatmo devices and data like weather station or camera data from Python 3.
 For more detailed information see [dev.netatmo.com](http://dev.netatmo.com)
 
 This project has no relation with the Netatmo company.
 
-Install
--------
+## Install
 
 To install pyatmo simply run:
 
@@ -31,14 +28,12 @@ Once installed you can simply add `pyatmo` to your Python 3 scripts by including
 
     import pyatmo
 
-Note
-----
+## Note
 
-The module requires a valid user account and a registered application. See [usage.md](./usage.md) for further information.
+The module requires a valid user account and a registered application.
 Be aware that the module may stop working if Netatmo decides to change their API.
 
-Development
------------
+## Development
 
 Clone the repo and install dependencies:
 
@@ -51,8 +46,7 @@ To add the pre-commit hook to your environment run:
     pip install pre-commit
     pre-commit install
 
-Testing
--------
+## Testing
 
 To run the full suite simply run the following command from within the virtual environment:
 
@@ -64,7 +58,7 @@ or
 
 To generate code coverage xml (e.g. for use in VSCode) run
 
-    python -m pytest --cov-report xml:cov.xml --cov smart_home --cov-append tests/
+    python -m pytest --cov-report xml:cov.xml --cov pyatmo --cov-append tests/
 
 Another way to run the tests is by using `tox`. This runs the tests against the installed package and multiple versions of python.
