+++ setup.cfg
@@ -39,10 +39,6 @@ exclude = tests
 [options.package_data]
 pyatmo = py.typed
 
-[flake8]
-max-line-length = 88
-ignore = W503, E501
-
 [pep8]
 max-line-length = 88
 ignore = W503, E501
