+++ src/pyatmo/helpers.py
@@ -1,39 +1,13 @@
 """Collection of helper functions."""
 from __future__ import annotations
 
-from calendar import timegm
-from datetime import datetime, timezone
 import logging
-import time
-from typing import Any
+from typing import Any, cast
 
 from pyatmo.const import RawData
 from pyatmo.exceptions import NoDevice
 
 LOG: logging.Logger = logging.getLogger(__name__)
-
-
-def to_time_string(value: str) -> str:
-    """Convert epoch to time string."""
-
-    return (
-        datetime.fromtimestamp(int(value), tz=timezone.utc)
-        .isoformat(sep="_")
-        .split("+")[0]
-    )
-
-
-def to_epoch(value: str) -> int:
-    """Convert time string to epoch."""
-
-    return timegm(time.strptime(f"{value}GMT", "%Y-%m-%d_%H:%M:%S%Z"))
-
-
-def today_stamps() -> tuple[int, int]:
-    """Return today's start and end timestamps."""
-
-    today: int = timegm(time.strptime(time.strftime("%Y-%m-%d") + "GMT", "%Y-%m-%d%Z"))
-    return today, today + 3600 * 24
 
 
 def fix_id(raw_data: RawData) -> dict[str, Any]:
@@ -48,7 +22,7 @@ def fix_id(raw_data: RawData) -> dict[str, Any]:
         if "_id" not in station:
             continue
 
-        station["_id"] = station["_id"].replace(" ", "")
+        station["_id"] = cast(dict, station)["_id"].replace(" ", "")
 
         for module in station.get("modules", {}):
             module["_id"] = module["_id"].replace(" ", "")
@@ -57,24 +31,6 @@ def fix_id(raw_data: RawData) -> dict[str, Any]:
 
 
 def extract_raw_data(resp: Any, tag: str) -> dict[str, Any]:
-    """Extract raw data from server response."""
-    if (
-        resp is None
-        or "body" not in resp
-        or tag not in resp["body"]
-        or ("errors" in resp["body"] and "modules" not in resp["body"][tag])
-    ):
-        LOG.debug("Server response: %s", resp)
-        raise NoDevice("No device found, errors in response")
-
-    if not (raw_data := fix_id(resp["body"].get(tag))):
-        LOG.debug("Server response: %s", resp)
-        raise NoDevice("No device data available")
-
-    return raw_data
-
-
-def extract_raw_data_new(resp: Any, tag: str) -> dict[str, Any]:
     """Extract raw data from server response."""
     raw_data = {}
