+++ src/pyatmo/modules/__init__.py
@@ -9,6 +9,7 @@ from .bticino import (
     BNEU,
     BNFC,
     BNIL,
+    BNLD,
     BNMH,
     BNMS,
     BNSL,
@@ -81,23 +82,25 @@ from .smarther import BNS
 from .somfy import TPSRS
 
 __all__ = [
-    "BNMS",
-    "BNAS",
     "BNAB",
-    "BNMH",
-    "BNTH",
-    "BNFC",
-    "BNTR",
-    "BNXM",
+    "BNAS",
     "BNCS",
     "BNCX",
     "BNDL",
     "BNEU",
+    "BNFC",
+    "BNIL",
+    "BNLD",
+    "BNMH",
+    "BNMS",
     "BNS",
     "BNSL",
-    "BNIL",
+    "BNTH",
+    "BNTR",
+    "BNXM",
     "Camera",
     "Dimmer",
+    "EBU",
     "Location",
     "Module",
     "NACamDoorTag",
@@ -117,6 +120,8 @@ __all__ = [
     "NDB",
     "NHC",
     "NIS",
+    "NLAO",
+    "NLAS",
     "NLC",
     "NLD",
     "NLDD",
@@ -126,7 +131,9 @@ __all__ = [
     "NLFN",
     "NLG",
     "NLIS",
+    "NLJ",
     "NLL",
+    "NLLF",
     "NLLM",
     "NLLV",
     "NLM",
@@ -139,16 +146,13 @@ __all__ = [
     "NLPS",
     "NLPT",
     "NLT",
+    "NLTS",
     "NLUF",
     "NLUI",
-    "NLAO",
-    "NLLF",
-    "NLUO",
     "NLunknown",
+    "NLUO",
     "NLUP",
     "NLV",
-    "EBU",
-    "Z3L",
     "NOC",
     "NRV",
     "NSD",
@@ -159,7 +163,5 @@ __all__ = [
     "Shutter",
     "Switch",
     "TPSRS",
-    "NLAS",
-    "NLTS",
-    "NLJ",
+    "Z3L",
 ]
